copr: Add vectorize from_base64.
Address comment.
