+++ components/tidb_query_common/src/lib.rs
@@ -7,7 +7,6 @@ pub mod error;
 pub mod execute_stats;
 pub mod metrics;
 pub mod storage;
-pub mod string;
 pub mod util;
 
 pub use self::error::{Error, Result};
